- Remove "virtual" from "needs_2passes()" - Remove "c_str()" on "post_processing()" - Reduce "FastIpAddress.h" and rename to "ipaddress.h" - Remove original "IpAddress.h" - Move "sctp_sctphdr_t" struct to "packet.h" and Remove "SCTPdefs.h" file
